pyproject.toml
[tool.poetry]
name = "heat-parser"
version = "0.1.0"
description = ""
authors = ["andrejmaznin <andrejmaznin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
bs4 = "^0.0.1"
requests = "^2.27.1"
pymorphy2 = "^0.9.1"
python-dateutil = "^2.8.2"
phonenumbers = "^8.12.45"
pre-commit = "^2.17.0"
isort = "^5.10.1"
mypy = "^0.942"
redis-om = "^0.0.21"
address = "^0.1.1"
language = "^0.6"
country-converter = "^0.7.4"
nameparser = "^1.1.1"
langdetect = "^1.0.9"
app-state = "^0.5"
current = "^0.4"
fake-useragent = "^0.1.11"
pytest = "^7.1.1"
aiohttp = "^3.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
